CPP: add #warning and #error directives
I plan to move some headers, and I want to create fallback headers that
"succeed loudly" so that programmers can be warned to change their source
code to compensate. None of the retro preprocessors have that ability, but
GCC (and other compilers of the GCC family, like clang) added #warning and
#error directives to do just this sort of thing -- get the compiler to
complain on demand.

This commit adds these directives.