CPP: Debug code for #define
When compiling the C preprocessor, if you build with -DDEBUG, it gains a
debugging argument "-x" that you can use to dump the macro table to the
standard error output after the file has been processed.

I found it useful to add some code that does a similar thing as the program
is running, so you can see how the preprocessor is understanding the macro
symbols as they are defined.

If you don't want the debug info (you probably don't), just don't build cpp
with -DDEBUG (it's not the default).