CPP: #define revamp
This one's kinda big.

Short version, we're adding ISO C "stringizing" and token pasting to the
#define directive.

Long version, we're making the preprocessor able to parse directives in
(almost) the same way that an ANSI/ISO preprocessor does. For example, an
ANSI preprocessor allows comments anywhere inside a directive, and it does
not process arguments inside quoted string literals, so the following is
all correct for an ANSI or ISO C preprocessor:

# /* this comment will always be hidden from the compiler */
# /* */ define a b
#define foo(a) printf("this -> a <- is not interpreted\n", a);

As far as I know, none of this is supposed to work for a K&R preprocessor,
which leaves us with a dilemma: which semantics to implement?

By default, the dcpp preprocessor runs in "standard" mode. All of the above
stuff works. If you have code you can't, or don't want to, change to work
by the more modern cpp rules, you can pass the -k flag to dcc.

the -k flag kicks dcpp into "K&R" mode. In this mode, directives don't
understand comments, and strings don't prevent argument expansion.
Everything else should remain the same. You can still do stringification
using #, and token pasting using ##.

What are those, you ask?

When the C preprocessor stopped replacing args inside strings (all in all,
a good thing), it also lost the ability to make strings that _include_ the
contents of an argument. To compensate for this, when it sees a #
immediately before the name of an argument, the preprocessor expands that
argument inside a string literal. It even works appropriately if you use
quote characters, backslashes, etc. Note: this _only_ works with macro
arguments -- you can't use #foo unless you're in a macro with an argument
named 'foo'.

#define STRING(foo) #foo
myString = STRING(this is my text);

expands to

myString = "this is my text";

Now, token pasting is a thing that the K&R preprocessor just can't do at
all -- it's unable to merge two "things" into one. It works like this:

#define pasted A ## B

When this macro expands, any whitespace between A and B will be removed;
they're "pasted" together. A and B can be anything that can form a single
token. In this case, pasted will expand to AB. This is useful for
generating code that will declare variables, for example:

#define MAKEVAR(type,name,val) type name ## _ ## type = val

MAKEVAR(int,Foo,10); will expand to:

int Foo_int = 10;

Because dcpp is not a _real_ ISO preprocessor, it can't tell when two
tokens aren't valid together, so if you screw it up it'll be the compiler
that tells you.