DCC: output cleanups, add usage
We only want to output our version info (the "logo") a maximum of once per
run, and only if something out of the ordinary happens.

Added a short usage help, accessed with -? like most OS-9 programs.

Also made a tweak to the formatting of our output, to give more space on
the screen for the output of our subprograms. I think it looks cleaner.