CC09: Make cross-compiled doubles work
I finally got 'round to tracking down exactly what was causing doubles to
not work in the cross-compiler. I initially tried to use the US Geological
Survey's "libvaxdata" library to convert native floats and doubles into
the PDP-11 format for the compiler, but it turns out we don't actually use
PDP-11 format for floating point numbers!

We _do_ use a format that has exactly the same bits as PDP-11, but they're
in a completely different order (even ignoring the endian difference). So
I went back to the fadjust.c code from somewhere (David Breeding?). I
wasn't really satisfied with that code either, since it uses a lot of
magic numbers, some of which are actually wrong (but which usually cancel
out giving the right result). So I rewrote it, and while it isn't
necessarily perfectly optimal, I think it's very clear what's going on.

Now, the other (decompiled, "Climax") version of the compiler does double
conversion in a *very* strange way -- when the lexer finds a double, it
first generates a native one and converts it to a native "OS-9" double,
then holds onto that binary blob until it's time to output. It does the
same thing to^Wfor long and int variables, which is probably one of the
reasons that compiler is so very fragile. We're using a totally different
approach, keeping all numbers in the host machine's native format (so they
can be easily manipulated) until immediately before they are output into
the assembly file. It's a lot simpler, and so far it seems to be quite
reliable.

The practical upshot is that you can cross-compile code using floats and
doubles. In the future, we should be able to do constant expression folding
for such numbers the same way we do with int expressions (that is, the
expression "a=2+2+4;" is always simplified by the compiler to "a=8;", and
we may be able to do the same for "a=2.0+2.0+4;", turning it into "a=8.0;")