New header: <stdint.h>
This is a first pass for a standard-ish <stdint.h> file.

It declares typedefs for intN_t and uintN_t for each type.

Some types required by ISO C99 are missing, because we don't have them.
Some others, such as int_leastN_t, are present but are currently hidden
from the compiler because it can't deal with the length of their type
names.

<stdint.h> also declares a number of macros for testing the limits of the
various integral types, and creating objects of those types. Again, there
are certain limits we haven't transcended yet, like the fact that we can't
make a UINT32_C(12) macro that's guaranteed to expand to an unsigned long.

We define int{8,16,32,ptr,max}_t and uint{16,ptr,max}_t, so you should be
able to safely declare "int32_t myvar;"

[u]intptr_t is the same size as a pointer to any type.