endian.h -- forgot the OS-9000 endian macro names
OS-9000 is the version of OS-9 that was rewritten in C and is relatively
portable, running on both big- and little-endian machines. For some reason,
they use their own markers, different from those used by _anyone_ else.

Adding them to the <endian.h> header for compatibility with OS9K source.