COpt: configurable patterns
This is a big one. This version of the optimizer gets rid of the hard-
coded assembly recognition patterns, and replaces them with two
configurable pattern files. The format is as follows:

---->8---- cut here ----8<----
opcode argument opcode argument ...
opcode argument ...

---->8---- cut here ----8<----

The first line is the _search_ pattern, a sequence of instructions to
search for in the incoming assembly code. The second line is the
_replacement_ pattern, a sequence of instructions that the optimizer will
replace the search pattern with. The third line must be blank.

Comments begin with #, _must_ start on the first character of a line, and
are treated as if the lines they are on do not exist. That is, comment
lines don't count as blank, they don't count as anything. If you put a
comment after a replacement pattern line, you still need a blank line
following it, and so on.

Each mnemonic or operand in a pattern is separated by EXACTLY one space,
so in the sequence "clra  clrb ", clra is followed by two spaces, and clrb
is followed by one. Each of the instructions in this sequence does have an
operand, the empty string. Any more or fewer spaces will almost certainly
result in unexpected behavior.

There are a few meaningful codes you can use:

#<	in a search pattern, #< matches an immediate value that fits in an
	8-bit register (A, B, E, F)
?	Like in shell 'globbing', matches any one character.
*	Like in the shell, matches zero or more characters, greedily.
	If you want to match only non-empty strings, "?*" will do that.
Lookback matching (search patterns):
^0	Matches any whole opcode or operand ("don't care"). This is the
	same as using just a single '*' wildcard.
^1	Matches the previous instruction's equivalent	part: in
	"adda #2 suba ^1", ^1 matches only the string "#2".
^2	Does the same as ^1, but matches against the equivalent part from
	the second instruction previous.

In replacement patterns, the codes are a little different. Because
replacement happens "backwards", if the replacement pattern has fewer
instructions than the search pattern, lookback refers to the corresponding
component of the _search_ pattern. Thus, for the operand of the last
instruction in a replacement pattern, ^1 expands to the text of the last
operand in the search pattern, and ^2 expands to the text of the
second-to-last operand in that same search pattern. Since I have not
needed more than two lookback levels, that's all there are at this time,
and the optimizer will reject ^3 etc. There is no ^0 for replacements.

Replacement patterns have another new notation: \1 and \2 expand to the
_globbed portion_ (the part that any wildcards in the pattern collected)
of the corresponding search pattern. That is, in the pattern:

lea? ?*,y
ld\1 >\1

when processing the opcode, \1 might contain the string "x", while when
processing the operand, \1 might contain the string "1104" or a label.

No checking is done to ensure	that lookback patterns have something valid
to look back at. Don't do it wrong. ;)

OK, now why did I do this? Two reasons -- first, to enable more, and more
complex, patterns, and second, so that you could control which
optimizations are done (specifically, to enable optimizations for OS-9
Level 2 and eventually Hitachi 6309 chips).

With the original hardcoded patterns, we were bumping up against limits in
rlink -- it couldn't handle the large structures we were trying to feed it.
With this new setup, the only limit we have for pattern space is available
memory (which we are in no danger of running out of any time soon).

Losing that limitation means I get to add more patterns, like the "C no Y"
optimization that can speed up programs on (Nitr)OS-9 Level 2 systems.

CnoY works by taking advantage of a difference between the Level 1 and
Level 2 kernels. On all OS-9, systems, when a program starts the U register
always points to the start of the program's private memory area.

On Level 1 that pointer can be almost anything, since the kernel and all
programs share a single pool of 64K. But on Level 2, the kernel gives each
program its own address space of up to 64K all to itself, and sets the U
register to zero accordingly.

Now, the C runtime system switches U and Y, so in a C program Y is the
start of the data area, so any code that uses global data will do
something like ldd myvar,y to get the value of "myvar" into D. This is
fine, but it's at least one cycle faster to do ldd >myvar instead.

TL;DR version -- if your program is only intended to work on a CoCo3, run
the optimizer with the -2 option, and your program will run a bit faster.

:)