COpt: cross-compile stuff
The variant structs (that is, a union comprised of two different struct
types, that are then used without giving their struct names) used in the
optimizer give modern C compilers a conniption. However, there is a
solution to be found in C11: anonymous unions. Anonymous unions allow you
to declare parts of structures to be variant without having to name the
union itself (or using a generic pointer, etc). Basically, it finds a way
to bless the sort of funny games people played with structures in
pre-Standard C. :)

Also, did some minor #define cleanups and increased the maximum number of
instructions considered when reordering code to optimize branch length,
when not running on OS-9 itself (possibly resulting in slightly more
efficient branching, at the expense of an insignificant increase of CPU
time on a modern machine).