Help: New "mouse" feature for NitrOS-9 Ease of Use
L. Curtis Boyle asked me to make Help's internal pager work in a similar
way to (Nitr)OS-9's terminal driver's "pause" functionality, and after
some discussions, I have agreed that it's worth doing in one specific way.

When the "pause" option is set for an SCF device (terminal), the terminal
driver basically counts how many "writeln" calls are made to that terminal,
and whenever that number reaches the number of lines the SCF device
descriptor thinks it has, the driver cuts off output until there is some
input. On input, the number resets and the driver starts counting writeln
calls again.

In Multi-Vue, the graphical shell's HELP support essentially depended on
this state of affairs to avoid overflowing the page length for long help
texts. Our new help program doesn't have that problem, because the new one
has a built-in pager that can tell the size of the screen, but it does lack
something the SCF driver does do -- SCF stops blocking output as soon as it
gets _any_ input, including joystick/mouse buttons on its window.

OK, now we do too. Sort of. We don't accept _any_ input, but we do now
treat a mouse/joystick button click exactly as if the spacebar were
pressed. To do this, we go to sleep and register for signals instead of
just a blocking read() call on stdin. When input is ready, the kernel
wakes us up and we do our thing pretty much just like before.