Help: Add a "no-ask" mode
This is another EOU thing, Curtis wanted the program to not be interactive
when run from the GUI, but I thought that was a bad idea, so we came up
with a compromise.

When called with the argument -n (for "no-ask"), any time Help would print
the top-level prompt, it instead exits. Nothing changes for subtopic
prompts -- if you "help -n shell", you still get a list of Shell subtopics
after the top-level help has displayed, and if you Enter from there you get
to the top-level prompt...where, instead of displaying that prompt, the
program exits back to the program that forked it.

This mode helps a bit for programs that might want to employ Help for a
program's internal help files.

Possibly for extra fun, just typing "help -n" gives you a list of top-level
help topics and exits immediately.