DCC: sprintf() is a thing
We were already using fprintf() for other stuff, so we really didn't need
to be repeatedly calling strcat() etc. to assemble strings from parts.

Let's just use single calls to sprintf() to build those strings.

We're also changing the order we're outputting child arguments in, to put
the output (-o=file) at the beginning for greater clarity.

And yes, all this makes the binary smaller too, which may or may not be
the real reason... :)

Binary regenerated too.