CC09: (experimental) unsigned char support
This adds an internal type UCHAR, which corresponds to the C type
"unsigned char". Consider it experimental -- it works for me, but there
may be hidden dragons I have not yet found. It's remarkably small, only
adding about 275 bytes to the binary.

At this point, it seems to be fully functional; you can cast to it, pass
values and variables to functions, etc.

Bang on it, will ya?