CC09: No way is this right...
This could be a major bug. the expression optimizer, when grabbing memory
for a double variable, sometimes actually wrote the pointer address to the
memory _pointed at_ by the dblptr pointer variable, instead of to the
variable itself. This is how the variable gets initialized, so if the value
of dblptr happens to be NULL or some other random value (since it's
allocated on the stack), that means randomly scribbling over memory. That's
some seriously bad juju.

Fixed.