CC09: make allocators more easily settable
lex.c contains the compiler's global memory allocator, grab(). We don't
ever return/free memory, and are very reluctant to allocate any.

The basic system-level allocation scheme used is based on the Unix sbrk()
"set break" system call, which increases the program's memory allocation
by N bytes, and returns a pointer to the start of the new space. OS-9 C
does support this call, but due to the unique way OS-9 handles memory this
is *not* the most efficient possible way to do it. It's mostly the safest,
though. When using sbrk (or malloc/free), you want to limit (if possible)
the memory you give the program to the amount of stack it needs, to
maximize what's left for dynamic allocations.

The most efficient way is a similar call, ibrk(). ibrk requests a pointer
within the program's initial allocation, so while the stack is growing
from the top of available memory, variables are growing from the bottom.
When using ibrk, you want to set your allocation as large as possible.

ibrk is somewhat less safe, because the stack and dynamic allocations are
sharing the same memory!

Anyway, it turns out that using ibrk() is the only way to get the Ultimuse
MIDI stuff compiled on a CoCo; it makes the compiler need a lot of stack,
but doesn't cause the compiler to do lots of memory requests, so declaring
we need All The Memory in the program header and using ibrk() to chop it up
into tiny bits looks like the way to go.

But since other systems don't have (or need) this scheme, we define a
symbol in cj.h, ALLOCATE, to be the name of a function that behaves like
sbrk(), in that when it fails it returns the value (char *) -1, not NULL.