CC09: unsigned long support
This commit adds support for "unsigned long" to the compiler, so now we
have unsigned versions of each integral type the compiler supports.

I initially thought we would need library support for unsigned long
compares (and Walter started writing assembly code for a _ulcmpr
function), but it turns out the original _lcmpr function should be
perfect for our purposes, since what it does is set the 6809's flags to
what they would be if the 6809 actually _had_ a 32-bit compare
instruction. So the unsigned branch instructions still work with long
integers the same way they work with smaller ints.

Walter had some trouble with conflicts, and I had some changes I wanted
to make, so this isn't exactly the same patch as what he wrote.

Signed-off-by: Jeff Teunissen <deek@d2dc.net>