CC09: more magic numbers going away
No real changes in the code, this just helps understanding a little.

In particular, the getcon(int *p, int n, int f) and defcon(same args)
functions take a number and a flag. The second argument, n, is the number
of _words_, not bytes, used by the pointer in the first argument, p. The
last argument, f, tells defcon() that *p is in a floating-point format.
It's not used on 6x09, but on IEEE machines f triggers conversion from the
native floating-point format to that of the compiler.