AR: Kreider Ar, version 1.2
As far as I am aware, this is the earliest version of Carl Kreider's "ar"
still in existence.

Some disambiguation:

The first version of ar was written at Bell Labs for Unix version 6, by
Ken Thompson and Dennis M. Ritchie. Its job is to bunch groups of files
into a single "archive" file. Notably, the format of archives made and
extracted by ar is not, and has never been, standardized (though there is
a "common" format used today by _most_ modern Unix systems).

In 1975, Brian W. Kernighan and PJ Plauger released _Software Tools_,
a textbook attempting to teach how to write good software. The language
used in Software Tools was RATFOR, a language designed by Kernighan which
looks a lot like C but is actually translated to FORTRAN 66.

In the early 1980s, Carl wrote an OS-9 version of ar in C, based on the
Ratfor version in _Software Tools_.

From what I do know of Carl's history, I can guess that version 1.0 was
_probably_ written for the OS-9 Level 1 system he built himself, version
1.1 was for the Level 2 upgrade, and version 1.2 was an OS-9/68000 port.