AR: VAX/VMS port, by Greg Walker
This is checked in on a branch to prevent later developments from breaking
it.

Some time after 1986, Greg Walker hacked Ar 1.2 to work on VAX/VMS as well
as on OS-9 (he tested it on a CoCo 3 running Level 2, so it must have been
1986+)

There were formatting issues with this, mostly a bunch of missing empty
lines (this is not uncommon with text files sent between CR and CRLF
operating systems).

I fixed one bug, in line 246 of ar.c. Greg made a mistake disabling an
errno check for error 214, he meant to disable it for VAX but instead
enabled it _only_ for VAX. I changed the #ifdef VAX to #ifndef VAX.