CC09: output formatting updates
These changes are to modernize the format of the compiler's diagnostic
output, to make it easier for external programs to parse, and to shrink
the size of the binary (it's getting big enough that that's a concern).

Even so, there's a new feature. :)

The function eputs() now translates tabs in its input into spaces, to
prevent interpretation by the terminal. The practical upshot of this is
that on CoCo (Nitr)OS-9, programs that use tabs to indent source code can
still properly display error messages.

The reason this is necessary is pretty stupid -- the GrfDrv terminal
driver controlling the display interprets the tab character not as a "tab"
command to move to the next tab stop, but as a command to move the cursor
up one line. Someone wasn't thinking too hard when they did that, but
now it's something we need to deal with.