Grep: Imported from Research Unix
This is the source to the 32V version of the Unix grep/egrep/fgrep
commands, which are extremely useful for searching for patterns in files.
So useful are they, that even today grep is an essential tool for any
system administrator. Its name comes from a command in the "ed" editor,
"g/re/p", which globally searches for a regular expression, and prints the
lines that match...which is what grep does for files.

grep itself uses "basic" regular expressions (BREs), which are identical
to the patterns used by "ed" or "sed", so some of the special characters
like (), {}, [], + and - must be escaped using backslash \ to have their
special meanings. '\(f\|c\)ake' matches both "fake" and "cake"

egrep supports "Enhanced" regular expressions (EREs). These are like BREs,
except the special characters need to be escaped if you _don't_ want them
to be processed. '(f|c)ake' matches both "fake" and "cake".

Of the two, egrep is the one that's more consistent, since a few of the
special characters (like ^, which matches the beginning of a line) work
identically in both versions.

fgrep does the searching, but it is only able to deal with normal strings.
It can't do wildcards or any special pattern matching, but it's faster
than grep or egrep.

These aren't ported to OS-9 yet, this is just the initial source code drop.