Greps: save some memory
OS-9 has a native sector size of 256 bytes, and no read of >256 bytes will
ever return more than that. Original grep is seriously old-school, and
doesn't use any read buffering, but egrep and fgrep do.

Set the sizes of those buffers to 256 bytes instead of Unix's 512.

Also, set the size of fgrep's string list to save a block of RAM for each
running copy. Since fgrep's binary is smaller than 8K this isn't critical,
but if it went over 8K it might not run any more. It's not that big of a
deal, it just means that your string list limit is 394 instead of 400 in
the VAX version.