CC09: clean up tokens (fixes a bug)
I am not _entirely_ sure why this fixes a bug dealing with the calling of
functions that have been declared to return signed long ints, but my
working hypothesis is that the compiler sometimes thought a function
returning long was actually a keyword.

There are a number of known constraints, some explicit and some implicit,
on the numbers that correspond to tokens:

* If it's a number that represents a character in the character set, its
  number must be below 128 (it has to be positive when in a signed char)
* The assignment operator tokens (ASSPLUS and friends) must be the tokens
  with the highest numbers.
* The PTREE parse-tree printing code uses an array 200 items long, so no
  token may have a number above 199.
* The numbers 0 through 15 are reserved for C basic types.
* The numbers 0x10, 0x20, and 0x30 are used as part of the type system,
  and they mean "pointer to", "array of", and "function returning"
  respectively. So a function returning int has the type 0x31.

The practical upshot of this is that a number of ranges may potentially be
dangerous to use for tokens. Initially, there were seemingly random gaps
in the token list, and when I renamed a few of them I may have triggered
the bug. In this case, the type for "pointer to long" had the value 51,
or hex 0x36, which also mapped to the token KEYWORD and that may be what
broke the compiler when trying to call a function that had previously been
declared to return "signed long int", "long int", or just "long".

My first stage fix is just to move all tokens out of the 0x30..0x3f range,
and push the token space all the way out to the full range 0..199. This
may not be good enough, because it may be _possible_ for arrays and
pointers to also conflict.

But let's handle one thing at a time, shall we?