New header: <getopt.h>
The Kreider libc has a Unix-compatible getopt() function. At some point,
someone made a header for it and called it <arg.h>, but nothing else uses
that name anywhere and they forgot to document anything. The public domain
DECUS version of getopt() used the header <getopt.h>, and so does the
modern GNU getopt_long() function, so we will follow suit.

The UNIX function is declared in the header <unistd.h>, and I have made
one, but it's not quite ready for prime time yet.