CC09: make our new types work in structs
It turns out that adding the types to the compiler as a whole doesn't make
them work everywhere, because the lexer has a special "struct mode".

Inside a struct or union, the compiler has an entirely different set of
keywords it uses, and adding keywords to the main set has no effect on how
structs are dealt with. That is, "continue" is a keyword _except_ when the
compiler is parsing the inside of a struct or union.

So, let's add our new types to the "mos" keyword table, shall we? This
makes our new types work in structs.