CC09: unsigned fixes
We missed a case in function definition, where an argument declared as an
unsigned long int looks like it was being assigned a size of 2 instead of
4 bytes.

Also, when converting to unsigned char, convert to unsigned int first
instead of just int. Since we just chop off the high bits when converting
to uchar anyway, let's give that conversion the best possible chance to
get the best result.