CC09: enable FUNCNAME for cross-build
The compiler has the ability to track where in the source file the current
function is found. The FUNCNAME symbol currently enables some extra output
in the assembly, turning 'ttl' lines from:

 ttl funcname

into:

 ttl funcname,filename,line

I do have a few ideas about possibly adding __func__ (a C99+ feature that
expands to the full name of the current function) to the compiler, but
this doesn't do that. Maybe later.