DCC warnings, step 1
To have proper warnings, we need to be able to signal them from the
compiler _or_ the preprocessor without them getting flagged as errors --
we currently have no way to do that. Also, the existant error output from
the preprocessor seems to have been going to stdout instead of stderr...not
good.

I'm adding a command code for dcpp to use that the compiler will recognize
as a warning message. Same as an error, only it's not flagged as one --
the message is just printed to stderr with the location sent from dcpp.

Actually, this is kind of a dual update -- previously, errors and warnings
sent from the preprocessor weren't actually flagged as such unless they
were "fatal" -- in which case the compiler would abort as soon as one was
seen. Not ideal. But now the preprocessor will be able to report warnings,
errors, and fatal errors and the compiler should know what to do.

Also, nuking the pfile() function that was only called from one place in
the middle of some other printfs. Very silly.