CPP: multi-compile for new/old compilers
The native build of the preprocessor now outputs both 'cpp' and 'dcpp'
binaries, to be used with the old McCosh/Microware compiler and with DCC.

Why? Well, it has to do with c.prep, the McCosh compiler's terrible, awful,
no good, really bad preprocessor.

Even in 1982, when it was written, c.prep was not up to scratch with the C
of the time -- which already had #if, conditional expressions, etc. in
addition to #if[n]def, etc. c.prep doesn't have those (or much of anything
else), so if we want to modernize the headers like I've been doing, we need
to replace it.

Many people over the years have looked to McDowall's "ANSI-like" C_PREP
preprocessor, and in many cases it works all right, but it's significantly
slower than c.prep _and_ it improperly breaks on a nontrivial amount of
source code. It also uses fixed-length arrays for a lot of stuff internally,
so it runs out of memory on a regular basis and can't be extended.

Enter cpp/dcpp -- I have yet to find a single piece of code that this thing
chokes on that it shouldn't, and it's way faster than c_prep -- almost as
fast as the original c.prep.

When built as cpp (that is, when DCC is not defined in the build), the
preprocessor is always in "K&R" mode, and the only macros it pre-defines
are _OS9 and OS9. As such, it should be an excellent replacement for c.prep
without breaking code that builds with it (while not slowing down your
builds or running out of memory seemingly at random).

And importantly for me, I can actually un-suckify the headers without
breaking the old compiler. :)

(expect there to be changes soon to cc, cc1, cc2 to make them use cpp)