CPP: use all memory
I finally found a program that tested the limits of (d)cpp -- UltiMusE 3.
And boy, did it! I was able to get it working by adding a bunch of memory
pages to the initial reservation (for stack space), but that would cause
an unacceptable level of wasted stack space in the general case, meaning
fewer macros available to programmers, and I don't want that. I want to be
able to use as much stack as I need, _or_ use as many macros as I want.

It seemed like the OS-9 C architecture was conspiring to keep me from being
able to supply the best of both worlds, until I found ibrk().

In Unix (and OS-9) C the low-level memory allocator that forms the basis
for malloc() is sbrk() (short for "set break" which makes sense on old Unix
but nowhere else).

On OS-9, calling sbrk() with a positive number means you are asking for a
pointer to a memory location, and possibly requesting "new-to-you" memory
to be added to your process space. Once you run out of free space, sbrk()
fails and you don't get more memory. Meanwhile, your initial memory
allocation is used for your stack and pretty much _only_ your stack. Adding
more memory using the shell's #nnK stuff doesn't give C programs more
memory, it actually takes memory away from sbrk (and thus malloc), making
your program harder to run.

HOWEVER... ibrk() is able to come to the rescue. ibrk() works just like
sbrk(), only instead of requesting more memory from OS-9, it just doles
out pointers from within the initial allocation. This can suck too, because
you're passing out pointers to the same region of memory your stack comes
from, and that's usually pretty small, so it's not unlikely that you're
shooting yourself in the foot...

...unless you just ask for All The RAM in your initial allocation. So,
let's do that. We're using ibrk() to get our memory, and making the linker
give (d)cpp 40K of initial memory, shared between stack (which grows from
the top of memory) and heap (which grows from the bottom of memory). If
they meet, there will be a big boom, so try not to use lots and lots of
defines in very large programs?

Also, #ifdef out the odd-size memory request block -- it should only affect
word-addressed machines.