(D)CPP: short circuit evaluation
CPP has the same operator precedence as C, so a conditional expression like

#if 0 && foo

or

#if (1 || bar)

can be fully determined by the part of the expression to the left of the
operator, and when that occurs the right side of the expression isn't
considered. All the stuff on the right side must still be _legal_, so it
isn't completely ignored, but its value doesn't matter (because the
expression is already known to be true/false as a whole). This property of
logical operators is used a lot in expressions such as
"#if defined(VALUE) && VALUE == 2"; If we didn't check that it was defined
first, we might get a warning that VALUE is being used without being
defined. But since the operators "short circuit", we get to avoid this
warning.

But that's where _our_ preprocessor comes in and pees in our breakfast
cereal. Our preprocessor does its evaluation at the same time as parsing,
so it can't tell the difference between "left side" and "right side"
without some help.

Well, here's some help. :) We can now tell when the expression has already
been decided, so we now know to suppress warnings that depend on the value
of the stuff to the left of the boolean operator.

For GitHub: (Closes #6)