CC09: cleanup/explain some numbers
This (partially) reverts commit 19f7ea6578b069861cac9c6bcdecfdeefd81737a.

I was a little too zealous getting rid of numbers in codgen.c; I failed to
read the code closely enough. Where the code is unstacking an object 2 bytes
in size (an int or pointer, usually), it emits code to increment the 6x09
stack register S by two bytes, and sets the SP variable accordingly. The
original code _correctly_ used bare numbers for this case, since the
generated assembly includes specific changes to the S register.

Obviously, if we actually changed the size of int, we would have to change
all of these bare numbers, but we'd also have to change the generated
code--and without the bare numbers, we'd have no obvious way to tell.
So, this returns the numbers to their rightful places.