CC09: fix switch() dropping a constant
Among other programs, Awk does something fairly weird with switch
expressions: it does "switch (yybgin-yysvec-1) { ... }" which is some next-
level weird crap involving pointer arithmetic, but it _works_. Or at least,
it's supposed to.

Instead, the compiler calculates the difference between yybgin and yysvec,
and just _throws away_ the constant 1 it's supposed to be subtracting. It's
apparently broken in every version of our compiler all the way back to at
least the Tandy release. Until now. :)

What happened was ldxexp() was failing to generate code to apply the offset
that had already been calculated by tranxexp() and put into the expression
tree. A single-line change takes care of it, and a bug that's been lurking
in the compiler for 39 years bites the dust.

For GitHub: (Closes #14)