KLibc: build system setup
Before we can start fixing bugs in the C library, we first need to be able
to rebuild it. Now, at this time we can't use DCC to rebuild klibc, because
DCC is using RMA as its assembler instead of the old c.asm.

c.asm is actually an older version of RMA, and it produces a slightly
different version of Relocatable Object Format (ROF) files that the linker
(rlink) puts together into binaries. RMA produces version 1 ROF, while
c.asm produces version 0 ROF.

The two versions appear to be identical, except version 1 files have an
extra "common section" at the end. In almost all version 1 files, this last
section is empty, so it manifests as two zero bytes at the end.

rlink is able to process these files, but for some reason it's NOT able
to process them when they're combined into a library. It doesn't read the
version number of the object and use that to trigger checking for the
common section, so as a result we can't currently use version 1 ROF files
bundled into a library...so we need to use c.asm to assemble code for libs.

David Breeding's "R63" modification of RMA is able to force version 0
output, so some time before I start making a 6309 version of the compiler I
will need to switch to that assembler, but I want to study it first and try
to wrangle its code into some nicer shape before I make it a critical piece
of compiler infrastructure.

I also want to make a linker that's able to handle both ROF versions. I
believe our current 'rdump' tool is now able to handle version 1 objects in
libraries, I should just need to port that code into rlink as well.