KLibc: fgetc()/fputc() "implemented"
The C standard defines the fgetc() and fputc() functions as being identical
to the original functions getc() and putc(), respectively, except it's
legal to implement getc/putc as macros. As it happened, our system did it
exactly backwards -- if you wanted the f* functions, you would have to
define them as macros, making them potentially dangerous to use.

This adds fgetc and fputc functions to the library by simply adding the
symbols to the code that was already there.

Also, includes my updated <stdio.h> without the f* macros.

For GitHub: (Closes #7)