KLibc: fix Y2K bug in asctime()
This is actually what made me get into the compiler and C library in the
first place.

Years ago, I was writing an 'ls' utility, and the dates of the files were
displayed as having been updated in the year 19114. Yeah, right. :)

asctime() now correctly uses %4d and adds 1900 to the OS-9 year to get the
real one, instead of "19%02d".

For GitHub: (Closes #3)