KLibc: fix _prgname()
When he made his OS-9/6809 C library, Carl implemented a lot of functions
from OS9/68000's C library. One of them was _errmsg(), a wrapper for printf
that returns a passed-in error code, so that you can do stuff like:

    exit(_errmsg(errno, "Couldn't create file %s\n", filename));

which should print "myprogram: Couldn't create file foo" and exit with the
OS-9 error code 214 or whatever.

But in a program I was testing, instead it seemed to be inserting a
character at the beginning of the output string, causing the program to
do weird things to the terminal. Well, it turned out that there wasn't a
bug in _errmsg(), but in _prgname() --  a simple function to return the
name of the program.

Well, not so simple. The actual code for _prgname has to locate the
secondary module name string, an undocumented feature of the linker that
puts a null-terminated copy of the name field in the OS-9 module header,
right at the end of the binary. The problem happens because _prgname does
it pretty blindly. It found the end of the program module, skipped over the
CRC field right at the end, and kept going backwards until it found a NUL.
Then the code would consider the following character to be the beginning of
the program's name.

Maybe that worked on Carl's home-built OS-9 machine, I don't know, but it
doesn't work with our toolchain. I don't think my solution is the best
possible solution (that would be to modify rlink to add a symbol pointing
directly to the secondary name string), but it does have the benefit of not
requiring changes in other software. :)

The new code still assumes the location of the secondary name string, but
it first counts the number of characters in the _primary_ name string, the
high-bit-terminated one hooked up to the actual module header. It then
calculates the correct address for the secondary name string and returns
that. Unless we really screw up rlink, it should be more reliable than the
old code.

Note: if you hack the name of a program by binary patching, the new code
will use the length of the new name and disagree with argv[0]. Try not to
do that. :)

Thanks to William "lostwizard" Astle for help improving the assembly I had
come up with. It's smaller and presumably a little faster.

Also, library files regenerated.

For GitHub: (Closes #2)